+++ scripts/destroy_resource_policies.py
@@ -17,7 +17,7 @@ import argparse
 import logging
 from pathlib import Path
 from suspend import batch_execute, truncate_iter, wait_for_operations
-from util import compute, config_root_logger, project, parse_self_link
+from util import lkp, compute, config_root_logger, parse_self_link
 
 logger_name = Path(__file__).name
 log = logging.getLogger(logger_name)
@@ -63,7 +63,7 @@ def main(args):
 
     result = (
         compute.resourcePolicies()
-        .aggregatedList(project=project, filter=filter)
+        .aggregatedList(project=lkp.project, filter=filter)
         .execute()
     )
 
@@ -97,12 +97,8 @@ if __name__ == "__main__":
 
     logfile = (Path(__file__).parent / logger_name).with_suffix(".log")
     if args.debug:
-        config_root_logger(
-            logger_name, level="DEBUG", util_level="DEBUG", logfile=logfile
-        )
+        config_root_logger(logger_name, level="DEBUG", logfile=logfile)
     else:
-        config_root_logger(
-            logger_name, level="INFO", util_level="ERROR", logfile=logfile
-        )
+        config_root_logger(logger_name, level="INFO", logfile=logfile)
 
     main(args)