+++ scripts/suspend.py
@@ -23,11 +23,14 @@ from pathlib import Path
 
 import util
 from util import (
+    groupby_unsorted,
+    log_api_request,
     run,
     execute_with_futures,
     batch_execute,
     ensure_execute,
     subscription_delete,
+    to_hostlist,
     wait_for_operations,
     separate,
     is_exclusive_node,
@@ -54,33 +57,40 @@ def truncate_iter(iterable, max_count):
 
 def delete_instance_request(instance, project=None, zone=None):
     project = project or lkp.project
-    return compute.instances().delete(
+    request = compute.instances().delete(
         project=project,
         zone=(zone or lkp.instance(instance).zone),
         instance=instance,
     )
+    log_api_request(request)
+    return request
 
 
 def delete_instances(instances):
     """Call regionInstances.bulkInsert to create instances"""
-    if len(instances) == 0:
-        return
     invalid, valid = separate(lambda inst: bool(lkp.instance(inst)), instances)
-    log.debug("instances do not exist: {}".format(",".join(invalid)))
+    if len(invalid) > 0:
+        log.debug("instances do not exist: {}".format(",".join(invalid)))
+    if len(valid) == 0:
+        log.debug("No instances to delete")
+        return
 
+    valid_hostlist = util.to_hostlist(valid)
     if lkp.cfg.enable_reconfigure:
-        count = len(instances)
-        hostlist = util.to_hostlist(valid)
-        log.info("delete {} subscriptions ({})".format(count, hostlist))
+        count = len(valid)
+        log.info("delete {} subscriptions ({})".format(count, valid_hostlist))
         execute_with_futures(subscription_delete, valid)
 
     requests = {inst: delete_instance_request(inst) for inst in valid}
+
+    log.info(f"delete {len(valid)} instances ({valid_hostlist})")
     done, failed = batch_execute(requests)
     if failed:
-        failed_nodes = [f"{n}: {e}" for n, (_, e) in failed.items()]
-        node_str = "\n".join(str(el) for el in truncate_iter(failed_nodes, 5))
-        log.error(f"some nodes failed to delete: {node_str}")
+        for err, nodes in groupby_unsorted(lambda n: failed[n][1], failed.keys()):
+            log.error(f"instances failed to delete: {err} ({to_hostlist(nodes)})")
     wait_for_operations(done.values())
+    # TODO do we need to check each operation for success? That is a lot more API calls
+    log.info(f"deleted {len(done)} instances {to_hostlist(done.keys())}")
 
 
 def suspend_nodes(nodelist):
@@ -110,6 +120,7 @@ def delete_placement_groups(job_id, region, partition_name):
     if failed:
         failed_pg = [f"{n}: {e}" for n, (_, e) in failed.items()]
         log.error(f"some nodes failed to delete: {failed_pg}")
+    log.info(f"deleted {len(done)} placement groups ({to_hostlist(done.keys())})")
 
 
 def epilog_suspend_nodes(nodelist, job_id):
@@ -119,6 +130,7 @@ def epilog_suspend_nodes(nodelist, job_id):
         nodes = util.to_hostnames(nodes)
     if any(not is_exclusive_node(node) for node in nodes):
         log.fatal(f"nodelist includes non-exclusive nodes: {nodelist}")
+        exit(1)
     # Mark nodes as off limits to new jobs while powering down.
     # Have to use "down" because it's the only, current, way to remove the
     # power_up flag from the node -- followed by a power_down -- if the
@@ -139,23 +151,35 @@ def epilog_suspend_nodes(nodelist, job_id):
 
 def main(nodelist, job_id):
     """main called when run as script"""
-    log.debug(f"main {nodelist} {job_id}")
+    if job_id is None:
+        log.debug(f"SuspendProgram {nodelist}")
+    else:
+        log.debug(f"EpilogSlurmctld exclusive suspend {nodelist} {job_id}")
     nodes = util.to_hostnames(nodelist)
 
     # Filter out nodes not in config.yaml
     cloud_nodes, local_nodes = lkp.filter_nodes(nodes)
-    log.debug(
-        f"Ignoring local nodes '{util.to_hostlist(local_nodes)}' from '{nodelist}'"
-    )
-    log.debug(f"Using cloud nodes '{util.to_hostlist(cloud_nodes)}' from '{nodelist}'")
+    if len(local_nodes) > 0:
+        log.debug(
+            f"Ignoring local nodes '{util.to_hostlist(local_nodes)}' from '{nodelist}'"
+        )
+    if len(cloud_nodes) > 0:
+        log.debug(
+            f"Using cloud nodes '{util.to_hostlist(cloud_nodes)}' from '{nodelist}'"
+        )
+    else:
+        log.debug("No cloud nodes to suspend")
+        return
     nodes = cloud_nodes
 
     if job_id is not None:
         _, exclusive = separate(is_exclusive_node, nodes)
-        if exclusive:
+        if len(exclusive) > 0:
             hostlist = util.to_hostlist(exclusive)
             log.info(f"epilog suspend {hostlist} job_id={job_id}")
             epilog_suspend_nodes(exclusive, job_id)
+        else:
+            log.debug("No exclusive nodes to suspend")
     else:
         # suspend is allowed to delete exclusive nodes
         log.info(f"suspend {nodelist}")
@@ -173,7 +197,19 @@ parser.add_argument(
     help="Optional job id for node list. Implies that PrologSlurmctld called program",
 )
 parser.add_argument(
-    "--debug", "-d", dest="debug", action="store_true", help="Enable debugging output"
+    "--debug",
+    "-d",
+    dest="loglevel",
+    action="store_const",
+    const=logging.DEBUG,
+    default=logging.INFO,
+    help="Enable debugging output",
+)
+parser.add_argument(
+    "--trace-api",
+    "-t",
+    action="store_true",
+    help="Enable detailed api request output",
 )
 
 
@@ -189,14 +225,13 @@ if __name__ == "__main__":
         args = parser.parse_args()
 
     util.chown_slurm(LOGFILE, mode=0o600)
-    if args.debug:
-        util.config_root_logger(
-            filename, level="DEBUG", util_level="DEBUG", logfile=LOGFILE
-        )
-    else:
-        util.config_root_logger(
-            filename, level="INFO", util_level="ERROR", logfile=LOGFILE
-        )
+
+    if cfg.enable_debug_logging:
+        args.loglevel = logging.DEBUG
+    if args.trace_api:
+        cfg.extra_logging_flags = list(cfg.extra_logging_flags)
+        cfg.extra_logging_flags.append("trace_api")
+    util.config_root_logger(filename, level=args.loglevel, logfile=LOGFILE)
     log = logging.getLogger(Path(__file__).name)
     sys.excepthook = util.handle_exception