+++ terraform/slurm_cluster/examples/slurm_controller_instance/simple/main.tf
@@ -38,6 +38,7 @@ module "slurm_partition0" {
       node_conf              = {}
 
       additional_disks         = []
+      bandwidth_tier           = "platform_default"
       can_ip_forward           = false
       disable_smt              = false
       disk_auto_delete         = true
@@ -85,6 +86,7 @@ module "slurm_partition1" {
       node_conf              = {}
 
       additional_disks       = []
+      bandwidth_tier         = "platform_default"
       can_ip_forward         = false
       disable_smt            = false
       disk_auto_delete       = true